max file size feature

+++ cmd/server/view/files.go
@@ -23,7 +23,7 @@ func Files(fs ifs.Filesystem, linkID string, renderDownloadLinks bool) ([]templa
 		fileViews = append(fileViews, template.FileView{
 			Name:         file.Name,
 			UploadedAt:   file.ModTime.Format(DateTimeFormat),
-			Size:         bytesToHuman(file.Size),
+			Size:         bytesToHuman(uint64(file.Size)),
 			DownloadLink: downloadLink,
 			DeleteLink:   fmt.Sprintf("/link/%s/file/%s/delete", linkID, file.Name),
 		})
